Reestructuración y saneamiento integral del proyecto
Principales cambios:
   * Arquitectura: Refactorización de los 7 microservicios bajo el patrón CSR (model, controller, service,
     repository).
   * Integración: Configuración de Eureka Client, API Gateway, Swagger/OpenAPI y estandarización de archivos
     .properties (dev/test).
   * Cumplimiento Rúbrica:
     - Corrección de relaciones JPA (@OneToMany/@JoinColumn) y Queries (JPQL/Native).
     - Implementación de Bean Validation en DTOs y persistencia de datos (Flyway, Liquibase, CommandLineRunner).
     - Ajuste de FeignClients para comunicación dinámica entre servicios.
   * Saneamiento: Resolución de errores de compilación, importaciones, tipos de datos y advertencias de Lombok
     (@Builder.Default).

+++ ms-vehiculos/src/main/java/com/example/ms_vehiculos/dto/VehiculosResponseDTO.java
@@ -17,7 +17,7 @@ public class VehiculosResponseDTO {
     private Integer id;//-- el cliente necesita ver su id
     private String patente;
     private String marca;
-    private Integer precioDiario;
+    private Double precioDiario;
     private boolean disponible;
     private LocalDate fechaIngreso;
 }
